data: widen token ids u16 -> u32 for large vocabularies
Qwen's 151936-entry vocabulary overflows u16. Widen the in-memory token
path to u32 while keeping the u16 on-disk format for existing small-vocab
(char/GPT-2) datasets:

- binio: add read/write_u32_bin + read_tokens_u32 (picks .u32.bin over
  .bin), Meta::stoi -> u32, Meta::vocab_only, itos optional in from_json.
- Tokenizer trait + CharTokenizer + Gpt2Bpe -> Vec<u32>; expose
  bytes_to_unicode/get_pairs/bpe_merge as pub(crate) for reuse.
- loader: TokenDataset/BatchConfig token fields -> u32.
- model::train::load reads via read_tokens_u32; eval/gpt_cli/bench drop
  now-redundant casts; prepare narrows char/GPT-2 ids back to u16 on write.

No behaviour change for existing datasets (all 40 data tests pass).

+++ crates/bench/src/model.rs
@@ -199,7 +199,7 @@ pub fn greedy_until_newline(
         let window: Vec<u32> =
             if ctx.len() > block { ctx[ctx.len() - block..].to_vec() } else { ctx.clone() };
         let next = scorer.predict_next(&window);
-        let ch = tok.decode(&[next as u16]);
+        let ch = tok.decode(&[next]);
         if ch == "\n" {
             break;
         }
